add tests for getRecursiveOptionValues
